Added ball and paddle rendering. Must work on resizing

+++ src/BallVisualComponent.cpp
@@ -1,5 +1,5 @@
 #include "BallVisualComponent.h"
-BallVisualComponent::BallVisualComponent(double radius)
+BallVisualComponent::BallVisualComponent(float radius)
 	:
 	radius(radius)
 {
